Only send FIFO commands if the named FIFO exists.

+++ xbmc/script.avr-control-client/default.py
@@ -23,9 +23,13 @@ assert len(sys.argv) > 0
 args = sys.argv[1:]
 
 if args and args[0] in cmd_map:
-#	print "Sending '%s' to %s" % (cmd_map[args[0]], avr_fifo)
-	f = open(avr_fifo, "w")
-	print >>f, cmd_map[args[0]]
-	f.close()
+	if os.path.exists(avr_fifo):
+#		print "Sending '%s' to %s" % (cmd_map[args[0]], avr_fifo)
+		f = open(avr_fifo, "w")
+		print >>f, cmd_map[args[0]]
+		f.close()
+	else:
+		print "Cannot forward command '%s': %s does not exist" % (
+			args[0], avr_fifo)
 else:
 	print "Unknown command '%s'" % (", ".join(args))
